begin setting up for reliefTechnique autocomplete
