+++ lib/components/CrisisPlan/reliefTechniqueCrisisCard.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:relieflink/utils/constants.dart';
+import 'package:relieflink/activities/activities.dart';
+
+Widget reliefTextInput(String label, String placeholder) {
+  return Padding(
+      padding: const EdgeInsets.only(left: 20, right: 20),
+      child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+        Text(label,
+            style: const TextStyle(
+              height: 0,
+              color: AppColors.font,
+              fontFamily: 'MainFont',
+              fontWeight: FontWeight.w900,
+              fontSize: 17,
+            )),
+        TextField(
+          onChanged: (value) => {placeholder = value},
+          style: const TextStyle(
+              color: AppColors.font,
+              fontFamily: 'MainFont',
+              fontWeight: FontWeight.w600,
+              fontSize: 18),
+          decoration: InputDecoration(
+            isDense: true,
+            hintText: placeholder,
+          ),
+        )
+      ]));
+}
